Never released PRI channels when using Busy() or Congestion() dialplan apps.
When the Busy() or Congestion() application is used towards ISDN (an ISDN
progress is sent), the responding ISDN Disconnect or Release may contain
the ISDN cause user busy or one of the congestion causes.  In chan_dahdi.c
these causes will only set the needbusy or needcongestion flags and not
activate the softhangup procedure.  Unfortunately only the latter can
interrupt the endless wait loop of Busy()/Congestion().

Result: PRI channels staying in state busy for the rest of asterisk life
or until the other end times out and forces the call to clear.

(in issue 0014292)
Reported by: tomaso
Patches:
      disc_rel_userbusy.patch uploaded by tomaso (license 564)
      (This patch is unrelated to the issue.)


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@224260 f38db490-d61c-443f-a65b-d21fe96a405b