Update documentation in sip.conf.sample.
Update the documentation in sip.conf.sample in order to make it more clear
that directmedia/canreinvite do not cause Asterisk to ignore reINVITEs. It
is only used to stop Asterisk from generating a reINVITE, but does not stop
it from accepting them if necessary.

(closes issue #15644)
Reported by: lmadsen

git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@226382 f38db490-d61c-443f-a65b-d21fe96a405b