Fix reading samples from format_mp3 after ast_seekstream/ast_tellstream.
There is a bug when using ast_seekstream/ast_tellstream with format_mp3 in that
the file read position is not reset before attempting to read samples.  So when
we seek to determine the maximum size of the file (as in res_agi's STREAM FILE)
we weren't then resetting the file pointer so that we could properly read
samples.  This patch addresses that (in a similar manner to format_wav.c).

(closes issue #15224)
Reported by: rbd
Patches:
      20091230_addons_1.4_issue15224.diff uploaded by seanbright (license 71)
Tested by: rbd, seanbright

Review: https://reviewboard.asterisk.org/r/453


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@238014 f38db490-d61c-443f-a65b-d21fe96a405b