Play correct prompt when voicemail store failure occurs after attempted forward.
If a user's mailbox was full and a message was attempted to be forwarded to
said box, warnings on the console would indicate failure. However, the played
prompt was that of success (vm-msgsaved). Now storage failure is taken into
account and the correct prompt (vm-mailboxfull) is played when appropriate.

ABE-2123
SWP-1262



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@258029 f38db490-d61c-443f-a65b-d21fe96a405b