fixes crash in chan_console
There is a race condition between console_hangup()
and start_stream().  It is possible for console_hangup()
to be called and then the stream thread to begin after the hangup.
To avoid this a check in start_stream() to make sure the pvt-owner
still exists while the pvt lock is held is made.  If the owner
is gone that means the channel hung up and start_stream should
be aborted.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@262236 f38db490-d61c-443f-a65b-d21fe96a405b