Use sigaction for signals which should persist past the initial trigger, not signal.
If you call signal() in a Solaris signal handler, instead of just resetting
the signal handler, it causes the signal to refire, because the signal is not
marked as handled prior to the signal handler being called.  This effectively
causes Solaris to immediately exceed the threadstack in recursive signal
handlers and crash.

(closes issue #17000)
 Reported by: rmcgilvr
 Patches: 
       20100526__issue17000.diff.txt uploaded by tilghman (license 14)
 Tested by: rmcgilvr


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@266142 f38db490-d61c-443f-a65b-d21fe96a405b