Fix MWI notification transmission problems over SIP.
MWI updates were not being sent if no messages were found in the event cache.
This was corrected since a phone may need to clear its MWI status configured
previously from another mailbox.

Upon module or sip reload, MWI updates could not be sent due to the sipsock
socket not being set early enough in reload_config. The code handling the
descriptor assignment and such has simply been moved before the call to
build_peer.

Issuing a sip reload cleared the IP address of the peer, but skipped checking
the database for registration information. The database is now checked both
for sip reload and actually reloading the module.

If a transmission occurs before the do_monitor thread has started, do not
attempt to send a signal to it.

(closes issue #17398)
Reported by: ip-rob


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@276571 f38db490-d61c-443f-a65b-d21fe96a405b