Add ability to configure the Max-Forwards header in the dialplan, as well as in sip.conf configuration for the channel and for devices.
The Max-Forwards header is used to prevent loops in a SIP network. Each intermediary,
like SIP proxys and SBCs, decrement this counter and detects when it reaches zero,
at which point the SIP request is nicely killed in a SIP-friendly way.

Review: https://reviewboard.asterisk.org/r/778/

Thanks to dvossel for the review and good advice.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@276951 f38db490-d61c-443f-a65b-d21fe96a405b