Fix inband DTMF detection on outgoing ISDN calls.
This is a regression from the sig_pri split from chan_dahdi.  When a call is
first initiated, the inband DTMF detector is not enabled if it's an outgoing
ISDN call.  However, it needs to be turned on once the media path starts up.
This handling was put back in the open_media() callback of chan_dahdi.  In
sig_pri, open_media() calls were added to a few places where it was needed,
including handling of PRI_EVENT_RINGING, PRI_EVENT_PROGRESS, and
PRI_EVENT_PROCEEDING.

Thanks to rmudgett for helping me with the patch!


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@279916 f38db490-d61c-443f-a65b-d21fe96a405b