Detect when libsrtp cannot be linked in a shared library
The libsrtp build system currently does not produce a shared library
or a static library compiled with -fPIC, so on 64-bit systems it is
possible that we will get a compile error if libsrtp is installed and
res_srtp is selected in menuselect.

This patch attempts to detect this situation and provide the user with
instructions to work around the problem.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@282200 f38db490-d61c-443f-a65b-d21fe96a405b