Fixed how Asterisk destroys a dialog on channel hangup before invite receives a response.
If an ast_channel with a SIP tech pvt hangs up before the sip dialog gets a response
to its outgoing INVITE, Asterisk used to pretend_ack the INVITE.  This is not rfc
compliant and results in confusion at the other endpoint.  sip_pretend_ack will ack
and remove all the packets in the retransmit queue.  This means that the INVITE will
stop retransmitting, and that any response to that INVITE that comes after the pretend_ack
occurs will be ignored.

Instead of faking any sort of acknowledgement for an outgoing INVITE during an internal
hangup, we should let the protocol stack process the INVITE transaction and terminate
the dialog properly.  This is achieved by setting the PENDING_BYE flag.  When this flag
is used, once the dialog proceeds to an escapable state the transaction will either be
canceled with a SIP_CANCEL or completed followed immediately by a BYE.  Attempting to do
this any other way is incorrect.  If the endpoint is not responding to the INVITE request,
the INVITE must continue to be retransmitted until it times out which will result in the
dialog being destroyed.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@283690 f38db490-d61c-443f-a65b-d21fe96a405b