During request to dialog matching, verify init_ruri is present before comparing.
During request to dialog matching, we attempt a best effort routine for fork
detection which requires several elements to be in place.  The dialog's
initial request uri is one of those elements.  Since it is best effort,
if the init_ruri is not present for some reason we can not proceed with that
routine.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@284561 f38db490-d61c-443f-a65b-d21fe96a405b