Disables auth_options_request option by default.
The auth_options_request option was created to do authentication
on OPTIONS request just like INVITES are done.  Since it has been
noted that some endpoints use OPTIONS requests as a way of qualifying
a peer and that a 401 authentication response could result in
interoperability issues, this option has been disabled by default.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@285006 f38db490-d61c-443f-a65b-d21fe96a405b