Initialize buffers in getvar and getvarfull.
Initialize the buffers used to hold the result from GET VARIABLE or
GET VARIABLE FULL.  The bug report shows func_read returning garbage in
the result.  It assumed that the buffer passed in was initialized, like many
other functions do.  In the more common code path (through the dialplan), it
is initialized, so just initialize it here too.

(closes issue #19050)
Reported by: johnz


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@314822 f38db490-d61c-443f-a65b-d21fe96a405b