Fix the bounds-checking code.
The code that set the bit within the select bitfield was correct, but the
bounds-checking code was not.  The change to that line uses the new _bitsize
macro for clarity.  Also, FD_ZERO macro did not zero-out anything but the
first word of the bitfield, so this could have caused problems with modules
using that macro with the expanded bitfield.

(closes issue #18773)
 Reported by: jamicque
 Patches: 
       20110423__issue18773.diff.txt uploaded by tilghman (license 14)
 Tested by: chris-mac


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@315501 f38db490-d61c-443f-a65b-d21fe96a405b