CCSS generic agent with POTS and ISDN phones fail caller busy call-back test.
If the following is true after a CCSS activation:
* The generic agent is for an analog phone or ISDN phone.  (Caller party)
* The called party becomes available.
* The caller party is not available.

When the caller party becomes available, the caller is not alerted to the
called party being available.  The generic agent still thinks the caller
is busy.

* Fixed the generic agent device state event subscription to look for all
device states that are considered available.

* Encapsulated the device state test for CCSS generic device available in
cc_generic_is_device_available().  Made the generic agent and monitor use
the new function instead of the manually coded inline equivalent.

JIRA AST-559
JIRA SWP-3462


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@319758 f38db490-d61c-443f-a65b-d21fe96a405b