Fixes issue with video and text not being reinvited correctly with directmedia
If a SDP does not modify the session, we ignore it.  However, we were defaulting
no text and video support to true before checking to see if the sdp modified
anything or not.  This would result in process_sdp ignoring an sdp but removing
video and text from the call during direct media reinvites.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@325151 f38db490-d61c-443f-a65b-d21fe96a405b