Remove unnecessary libpri dependency checks in the configure script.
Using the --with-pri option with the configure script generated an error
about not having PRI_L2_PERSISTENCE if you did not have the absolute
latest libpri SVN checkout installed.

The AST_EXT_LIB_SETUP_DEPENDENT macro in the configure.ac script seems to
be for libraries that are dependent upon other libraries and not
necessarily for optional/added features within a library.

(closes issue ASTERISK-18535)
Reported by: Michael Keuter


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@335911 f38db490-d61c-443f-a65b-d21fe96a405b