Fixup a race condition in res_fax.c where FAXOPT(gateway)=no will turn off the gateway but the framehook is not destroyed.
this problem happens when a gateway is attempted in the dialplan and
the device is not available i may want to do fax to mail in the server
it will not be allowed.

instead of checking only AST_FAX_TECH_GATEWAY also check gateway_id

Reverts 338904

Fix some white space.



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@338950 f38db490-d61c-443f-a65b-d21fe96a405b