Reverting revision 333265 due to component connection problems it introduces.
I'm going to attempt some generic res_jabber cleanup and come up with a new fix for this
problem, but first it seems prudent to remove this rather broad attempt to fix it and
instead approach this problem either from the same angle but looking only at canceling
(or possibly rescheduling) the send when we absolutely know it will cause a segfault 
or, if that can't be easily accomplished, strictly from the devstate side of things.
Also, I'm pretty sure a lot of the code in res_jabber isn't thread safe.

(issue ASTERISK-18626)
(issue ASTERISK-18078)



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@339297 f38db490-d61c-443f-a65b-d21fe96a405b