Fixed segfault on core stop gracefully.
There was an issue that the cap and confcap pointers for each line and device
were being memcpy'd so they all pointed to the same ast_format_cap. On
destroying, a segfault occured on the second call to the same struct.

skinny reload now works again as well.

Tested by snuff (in trunk) and myself. 


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@339681 f38db490-d61c-443f-a65b-d21fe96a405b