Add option to check state when state is unknown
r341486 reverts r325483 this is a rework of the patch.
optimize to minimize load.

add option check_state_unknown to control whether a member with unknown
device state is checked there is a small % chance that calls will be sent
to the member when they on a call.

app_queue will see a device with unknown state as available and does not 
try verify the state without this option enabled.

Review: https://reviewboard.asterisk.org/r/1535/



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@341580 f38db490-d61c-443f-a65b-d21fe96a405b