Fixes memory leak in message API.
The ast_msg_get_var function did not properly decrement
the ref count of the var it retrieves.  The way this is
implemented is a bit tricky, as we must decrement the var and then
return the var's value.  As long as the documentation for the
function is followed, this will not result in a dangling pointer as
the ast_msg structure owns its own reference to the var while it
exists in the var container.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@346349 f38db490-d61c-443f-a65b-d21fe96a405b