Added support for all slin formats to app_originate
Previously, app_originate could not originate a call into a non-8kHz conference
bridge as the formats for non-8kHz slin codecs were not applied to the created
channel.  This patch adds all of the formats by default, such that if a created
channel has a codec that supports a higher sampling rate, a translation path
can be built between it and other channels.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@348265 f38db490-d61c-443f-a65b-d21fe96a405b