Fix premature free'ing of the frame committed in r349608
Even though we set the frame to the ast_null_frame and return that,
the caller of the frame hook may still need the frame.  This now is
a bit more careful about when it frees the frame, i.e., only under
the same conditions that applied when we duplicated it in the first
place.

git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/10@349822 f38db490-d61c-443f-a65b-d21fe96a405b