Find even more network interfaces.
The previous change made the code look for emN and pciN in addition to what
it did originally, which was search for ethN.  However, it needed to be looking
for pciN#N, so that's what it does now.

This also moves the memset() to be before every ioctl().
........

Merged revisions 353175 from http://svn.asterisk.org/svn/asterisk/branches/1.8
........

Merged revisions 353176 from http://svn.asterisk.org/svn/asterisk/branches/10


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@353177 f38db490-d61c-443f-a65b-d21fe96a405b