Clear the high order bit from the destination call number before sending.
send_apathetic_reply takes the incoming frame's source call number as the
destination call number for the outgoing frame.  If the incoming frame was a
full frame, then the high order bit of the source call number is set and will be
interpreted as a retransmit when sent back out as the destination call number.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@355182 f38db490-d61c-443f-a65b-d21fe96a405b