Fix incorrect sizeof() usage in features.c.
This didn't actually result in a bug anywhere, luckily.  The only place
where the result of these memcpys was used is in app_dial, and the only
field that it read out of ast_call_feature was the first one, which is an
int, so these memcpys always copied just enough to avoid a problem.


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@359069 f38db490-d61c-443f-a65b-d21fe96a405b