Allow specifying a port number for the MySQL server.
This patch allows you to specify a port number for the MySQL server.
It's useful if a MySQL server is running on a non-standard port.
Even though this module is deprecated in favor of func_odbc, someone
asked for this feature and it seems pretty harmless to add.

It has been tested using a number of combinations of with/without a
port number specified in the dialplan and changing the port number
 for mysqld.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@370534 f38db490-d61c-443f-a65b-d21fe96a405b