Fix ast_event_new unit test.
One of my recent commits broke this test.  The error was:

[test_event.c:event_new_test:214]: Events expected to be identical
have different size: 69 != 59

The difference in size occurred because the first event had
the EID IE added to the event twice.  ast_event_new() now always
adds it automatically.  Previously it only added it if there
were no IEs specified, which was kind of weird.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@370541 f38db490-d61c-443f-a65b-d21fe96a405b