Correct documentation for the MeetMe x flag
The documentation for the x flag for MeetMe incorrectly described its
function as closing down the conference when the last marked user left.
It actually causes the users with that flag to leave the conference
when the last marked user exits. The functionality of this flag is not
changing.
........

Merged revisions 370985 from http://svn.asterisk.org/svn/asterisk/branches/1.8
........

Merged revisions 370986 from http://svn.asterisk.org/svn/asterisk/branches/10


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@370987 f38db490-d61c-443f-a65b-d21fe96a405b