Ensure "rules" is tab-completable for "queue show"
Previously, tabbing at the end of "queue show" produced a list of
available queues about which information could be shown, but did not
include an alternative command, "rules", to access information about
queue rules. The "rules" item should now be shown in the list of
tab-completable items.

(closes issue AST-958)
Reported-by: John Bigelow


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@372444 f38db490-d61c-443f-a65b-d21fe96a405b