chan_sip: Note change in behavior to how directmediapermit/deny ACL works
r366547 introduced a change to the directmedia ACL for chan_sip which
modified the behavior significantly. Prior to the patch, this option would
bridge peers with directmedia if a peer's IP address matched its own
directmedia ACL. After that patch, the peer would check the bridged peer's
ACL instead. This change has been present since 1.8.14.0. That patched failed
to document the change in Upgrade.txt, so this patch adds mention of that
change to UPGRADE.txt (UPGRADE-1.8.txt in newer branches)

(issue AST-876)



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@372471 f38db490-d61c-443f-a65b-d21fe96a405b