Ensure listed queues are not offered for completion
When using tab-completion for the list of queues on "queue reset stats"
or "queue reload {all|members|parameters|rules}", the tab-completion
listing for further queues erroneously listed queues that had already
been added to the list. The tab-completion listing now only displays
queues that are not already in the list.

(closes issue AST-963)
Reported-by: John Bigelow


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@372517 f38db490-d61c-443f-a65b-d21fe96a405b