Test for Asterisk Version info
Doxygen uses the ASTERISKVERSION as a sub header. If a SVN export is done and not .svn or .version file exists it defualts to UNKNOWN__and_probably_unsupported which is honest but not great for the online docs.  During the "make progdocs" I added a test for this and just warned and ommitted the version.

(issue ASTERISK-20259)


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@374949 f38db490-d61c-443f-a65b-d21fe96a405b