Adjust RTP instance's available_formats callback to return the correct type.
The RTP engine public function that gets the available formats expects a
format_t to be returned; however when calling into an RTP instance's
callback to get the available formats, the callback returned an int.

This never was noticed in Asterisk because the two RTP engines included
do not provide an available_formats callback.

This introduces an API change, and the proposal for this change was brought
up on the Asterisk developers mailing list [1]. There was no public objection
to this change, so it is now being put in.

(closes AST-1054)
reported by Doug Bailey

[1] http://lists.digium.com/pipermail/asterisk-dev/2012-December/058058.html



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.8@378147 f38db490-d61c-443f-a65b-d21fe96a405b